# Conflicts:
#	.vscode/launch.json
#	.vscode/settings.json
#	okr-web/src/app/check-ins/page.tsx

+++ okr-service/src/test/java/org/phc/templatejavabe/domain/service/CheckInServiceTest.java
@@ -67,7 +67,7 @@ class CheckInServiceTest {
         
         Objective objective = new Objective();
         objective.setId(objectiveId);
-        objective.setProgress(60);
+        objective.setProgress(BigDecimal.valueOf(60));
         
         List<KeyResult> keyResults = Arrays.asList(keyResult);
